+++ exporter/prometheusremotewriteexporter/exporter.go
@@ -35,7 +35,7 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/consumer/consumererror"
-	"go.opentelemetry.io/collector/consumer/pdata"
+	"go.opentelemetry.io/collector/model/pdata"
 )
 
 const maxBatchByteSize = 3000000
@@ -130,8 +130,8 @@ func (prwe *PRWExporter) PushMetrics(ctx context.Context, md pdata.Metrics) erro
 
 					// handle individual metric based on type
 					switch metric.DataType() {
-					case pdata.MetricDataTypeDoubleGauge:
-						dataPoints := metric.DoubleGauge().DataPoints()
+					case pdata.MetricDataTypeGauge:
+						dataPoints := metric.Gauge().DataPoints()
 						if err := prwe.addDoubleDataPointSlice(dataPoints, tsMap, resource, metric); err != nil {
 							dropped++
 							errs = append(errs, err)
@@ -142,8 +142,8 @@ func (prwe *PRWExporter) PushMetrics(ctx context.Context, md pdata.Metrics) erro
 							dropped++
 							errs = append(errs, err)
 						}
-					case pdata.MetricDataTypeDoubleSum:
-						dataPoints := metric.DoubleSum().DataPoints()
+					case pdata.MetricDataTypeSum:
+						dataPoints := metric.Sum().DataPoints()
 						if err := prwe.addDoubleDataPointSlice(dataPoints, tsMap, resource, metric); err != nil {
 							dropped++
 							errs = append(errs, err)